requirements.txt
openai
fastapi==0.128.0
passlib==1.7.4
protobuf==6.33.2
pydantic==2.12.5
python-dotenv==1.2.1
python-jose==3.5.0
SQLAlchemy==2.0.45
uvicorn==0.40.0
pymysql
cryptography
argon2-cffi
python-multipart
slowapi
redis
PyJWT
fastapi-cache2[redis]
captcha